Add rr init logic

+++ firmware/src/system/task/task_func.c
@@ -3,7 +3,7 @@
 void task1 (void) {
     for (;;) {
         GPIOA->BSRR = GPIO_BSRR_BS5;
-        for(uint16_t i = 0; i < 700000; i++) {}
+        for(uint32_t i = 0; i < 700000; i++) {}
 
     }
 }
